refactor rv adapter to run in host process
- split up adapter so that RV side doesn't need OpenTimelineIO to run.
